requirements.txt
gspread==6.2.0
oauth2client==4.1.3
vk-api==11.9.9
ok-api==1.0.1
python-telegram-bot==13.15
environs==14.1.1
google-auth==2.39.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.167.0